+++ packages/heq-server-redis/index.js
@@ -1,6 +1,6 @@
-const kefir = require('kefir');
-const redis = require('redis');
-const runLua = require('run-lua');
+const kefir = require("kefir");
+const redis = require("redis");
+const runLua = require("run-lua");
 
 const lua_commit = `
   local event = ARGV[1];
@@ -21,9 +21,12 @@ local to = tonumber(
   ARGV[2] or redis.call('get', KEYS[1])
 );
 local newArray = {};
+local event
 
 for id=from,to do
-  table.insert(newArray, {id, redis.call('hget', KEYS[2], id)});
+  if redis.call('hexists', KEYS[2], id) == 1 then
+    table.insert(newArray, {id, redis.call('hget', KEYS[2], id)});
+  end
 end
 
 return newArray
@@ -38,17 +41,15 @@ const lua_latest = `
   return {id, redis.call('hget', KEYS[2], id)};
 `;
 
-const adapter = ({ url, ns = 'local' }) => {
+const adapter = ({ url, ns = "local" }) => {
   const redisClient = redis.createClient(url);
   const subClient = redis.createClient(url);
 
   subClient.subscribe(`${ns}::events`);
 
-  const events$ = kefir.fromEvents(
-    subClient,
-    'message',
-    (channel, message) => message
-  );
+  const events$ = kefir.fromEvents(subClient, "message", (channel, message) => {
+    return message;
+  });
 
   const commit = async event => {
     delete event.id;
@@ -69,7 +70,7 @@ const adapter = ({ url, ns = 'local' }) => {
     if (id === 0) {
       return {
         id: 0,
-        type: '@@INIT',
+        type: "@@INIT",
       };
     }
 
@@ -86,7 +87,7 @@ const adapter = ({ url, ns = 'local' }) => {
     try {
       const argv = [String(from)];
 
-      if (typeof to !== 'undefined') {
+      if (typeof to !== "undefined") {
         argv.push(String(to));
       }
 
@@ -100,29 +101,79 @@ const adapter = ({ url, ns = 'local' }) => {
         id,
       }));
     } catch (ex) {
-      console.error(ex);
       return [];
     }
   };
-
-  const subscribe = () => ({
-    events$: events$.map(message => {
-      const rawId = message.split(':')[0];
-      const id = Number(rawId);
-      const rawEvent = message.slice(rawId.length + 1);
-
-      const event = JSON.parse(rawEvent);
-      event.id = id;
-      return event;
-    }),
-  });
 
   const destroy = () => {
     redisClient.quit();
     subClient.quit();
   };
 
-  return { commit, subscribe, query, destroy, getLatest };
+  async function* generate(from, max, time, includingTypes = []) {
+    let i = from;
+    let subscription;
+
+    const filter = includingTypes.length
+      ? event => includingTypes.includes(event.type)
+      : alwaysTrue;
+
+    try {
+      while (true) {
+        const to = i + max;
+        const batch = await query({ from: i, to });
+
+        const matched = batch.filter(filter);
+        if (matched.length) {
+          // console.log("from query", matched);
+          yield matched;
+        }
+
+        const lastEvent = last(batch);
+        if (lastEvent) {
+          i = last(batch).id;
+        } else {
+          break;
+        }
+        // await sleep(time);
+      }
+
+      const queue = [];
+
+      subscription = events$
+        .map(message => {
+          const rawId = message.split(":")[0];
+          const id = Number(rawId);
+          const rawEvent = message.slice(rawId.length + 1);
+
+          const event = JSON.parse(rawEvent);
+          event.id = id;
+          return event;
+        })
+        .filter(event => event.id > i)
+        .filter(filter)
+        .observe(event => {
+          queue.push(event);
+        });
+
+      while (true) {
+        if (queue.length) {
+          const buffer = [...queue];
+          queue.length = 0;
+          // console.log("from sub", buffer);
+          yield buffer;
+        }
+        await sleep(time);
+      }
+    } finally {
+      if (subscription) subscription.unsubscribe();
+    }
+  }
+
+  return { commit, generate, query, destroy, getLatest };
 };
 
 module.exports = adapter;
+const sleep = ms => new Promise(resolve => setTimeout(resolve, ms));
+const last = array => (array.length >= 1 ? array[array.length - 1] : null);
+const alwaysTrue = () => true;