+++ packages/heq-server/src/utils.js
@@ -12,28 +12,49 @@ const getNumberFromReq = ({ header, query }) => req => {
   }
 };
 
+function getArrayFromReq({ header, query }) {
+  return req => {
+    const fromHeaders = req.headers[header];
+
+    if (fromHeaders) {
+      return fromHeaders.split(",");
+    }
+
+    const fromQuery = req.query[query];
+    if (fromQuery) {
+      return fromQuery.split(",");
+    }
+
+    return [];
+  };
+}
+
 const getLastEventId = getNumberFromReq({
-  header: 'last-event-id',
-  query: 'lastEventId',
+  header: "last-event-id",
+  query: "lastEventId",
 });
 const getBurstCount = getNumberFromReq({
-  header: 'burst-count',
-  query: 'burstCount',
+  header: "burst-count",
+  query: "burstCount",
 });
 const getBurstTime = getNumberFromReq({
-  header: 'burst-time',
-  query: 'burstTime',
+  header: "burst-time",
+  query: "burstTime",
 });
 const getRetry = getNumberFromReq({
-  header: 'retry',
-  query: 'retry',
+  header: "retry",
+  query: "retry",
+});
+const getIncludes = getArrayFromReq({
+  header: "includes",
+  query: "includes",
 });
 
 const tryParse = raw => {
   try {
     return JSON.parse(raw);
   } catch (ex) {
-    return { type: '@@RAW', payload: raw };
+    return { type: "@@RAW", payload: raw };
   }
 };
 
@@ -41,4 +62,5 @@ exports.getLastEventId = getLastEventId;
 exports.getBurstCount = getBurstCount;
 exports.getBurstTime = getBurstTime;
 exports.getRetry = getRetry;
+exports.getIncludes = getIncludes;
 exports.tryParse = tryParse;