+++ packages/jerni/lib/mapEvents.js
@@ -1,14 +1,21 @@
-const arrify = require('arrify');
+const arrify = require("arrify");
 const identity = x => x;
 
 module.exports = rules => {
-  return event => {
-    if (typeof rules[event.type] !== 'function') {
+  const meta = {
+    includes: Object.keys(rules),
+  };
+
+  function transform(event) {
+    if (typeof rules[event.type] !== "function") {
       return [];
     }
 
     const ops = arrify(rules[event.type](event));
 
     return ops.filter(identity);
-  };
+  }
+
+  transform.meta = meta;
+  return transform;
 };