+++ packages/jerni/tests/makeServer.js
@@ -1,13 +1,13 @@
-const adapter = require('@heq/server-lokijs');
-const micro = require('micro');
-const factory = require('heq-server');
+const micro = require("micro");
+const factory = require("heq-server");
+const inMemoryAdapter = require("heq-server/src/adapters/in-memory");
 
 const makeServer = async ({ ns, port }) => {
-  const queue = await adapter({
+  const queue = await inMemoryAdapter({
     ns: ns,
   });
 
-  const service = await factory({ queue });
+  const service = await factory({});
   const server = micro(service);
 
   return new Promise((resolve, reject) => {