+++ packages/jerni-base/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.0"></a>
+# [0.1.0](https://github.com/tungv/jerni/compare/@jerni/base@0.0.2...@jerni/base@0.1.0) (2019-04-12)
+
+
+### Features
+
+* **store:** connection class will check meta of its models ([826d335](https://github.com/tungv/jerni/commit/826d335))
+
+
+
+
 <a name="0.0.2"></a>
 ## 0.0.2 (2018-09-03)