+++ packages/jerni/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.7.0"></a>
+# [0.7.0](https://github.com/tungv/jerni/compare/jerni@0.6.2...jerni@0.7.0) (2019-04-12)
+
+
+### Features
+
+* **jerni:** add auto includes extraction ([370e029](https://github.com/tungv/jerni/commit/370e029))
+* **store:** connection class will check meta of its models ([826d335](https://github.com/tungv/jerni/commit/826d335))
+
+
+
+
 <a name="0.6.1"></a>
 ## [0.6.1](https://github.com/tungv/jerni/compare/jerni@0.6.0...jerni@0.6.1) (2018-11-13)